Tweak protoIndex validation to check union member
Union rendering will change. Before, union could be rendered
as top level message with the `oneOf` inside of it. This is
not supported anymore. The union are now rendered within
the structure where they're used. To obtain a top-level
message with only the union (like before), just wrap it in a
structure with a single field.

This also adds a constraint on union usage where union have to
be used in ONE structure. If they are not used, they are not
rendered. If they are used in multiple places, smithy-translate
will fail to compile.