Remove defaultValue trait
As I was trying to write the documentation for the alloy#defaultValue
trait, I kept struggling to write it. This is because, now, in Smithy
2.0 we have `smithy.api#default`. So I'm trying to justify the need
for a custom `defaultValue` trait.

From looking at the history, there were a few things that made sense:

* Smithy 1 did not have a default trait
* `defaultValue` selector is stricter: only on available on structure members
* `defaultValue` can't be used in along with `@required`

At this point in time, I believe we can get rid of `@defaultValue` and
instead use `@default` from the Smithy prelude. We still keep the
`DefaultValue` hint that we use internally as we can do some processing
on the intermediate model before turning it into a Smithy model.