Ensure protoIndex are valid along with `protoInlinedOneOf`
Thanks to @Baccata, I found issues related to `protoIndex` consistency
when using union w/ or w/o `protoInlinedOneOf`

Basically it comes down to this:
1. if a field of structure has a @protoIndex, all other fieds should have it
2. this extends to union members if the union is annotated with
`protoInlinedOneOf`

This PR tweaks the algorithm so that if the union does not use
`protoInlinedOneOf`, then a structure member referencing it would
need `protoIndex` if other fields also have it. And if the union uses
the `protoInlinedOneOf`, then the members of the union are validated
with the structure's member for consistency.