Allow using JSON transformations in OpenAPI Config
Smithy has two methods for converting a model to an OpenAPI object. The
one currently being used will only apply a subset of the OpenAPI Config
transformations: in particular, it will not apply any transformations
that could result in an invalid OpenAPI JSON, such as JsonAdd or any
substitutions. This behaviour is surprising as it means valid JSON
transformations of the OpenAPI are also silently ignored. As we do not
need the OpenAPI object _as_ OpenAPI but only to render it as JSON, we
can use the converter's convertToNode method, which applies all changes
specified in the config.